Platform views shouldn't receive pointer events when not laid out (#172043)
This PR makes platform views ignore pointer events when they're not laid
out.

Platform views register themselves for global pointer events when their
RenderBox attaches:


https://github.com/flutter/flutter/blob/9c626d9f9afa6da96433e3bfc4895492f87e0009/packages/flutter/lib/src/rendering/platform_view.dart#L370-L380

But a RenderBox that is attached is not necessarily laid out. The
problem in the issue seemed to be that a page was created offscreen
(attached and not laid out), but it still received pointer events, and
this caused errors with using `size` before layout.

This PR simply ignores pointer events when not laid out.

Fixes https://github.com/flutter/flutter/issues/83481