Remove jetifier usages from framework and engine (#173459)
Justification this is safe: 
useAndroidX which I left in this pr was added in 2018 and is one of 2
flags that control what compile behavior is used as android migrated
from the support library to androidx.
We (flutter android) believe we have migrated all support library usages
to android x.
Android Studio does not automatically insert the Jetifier flag in newly
created projects, suggesting a move towards less reliance on it.

I think even if plugins were using the support libraries this pr would
not break them because the targets being modified are tests in this repo
and the engine build which only uses androidx dependencies.

## Pre-launch Checklist

- [x] I read the [Contributor Guide] and followed the process outlined
there for submitting PRs.
- [x] I read the [Tree Hygiene] wiki page, which explains my
responsibilities.
- [x] I read and followed the [Flutter Style Guide], including [Features
we expect every widget to implement].
- [x] I signed the [CLA].
- [x] I listed at least one issue that this PR fixes in the description
above.
- [x] I updated/added relevant documentation (doc comments with `///`).
- [ ] I added new tests to check the change I am making, or this PR is
[test-exempt].
- [x] I followed the [breaking change policy] and added [Data Driven
Fixes] where supported.
- [x] All existing and new tests are passing.