Flutter driver deserialization (#172927)
Adds more descriptive ArgumentErrors to flutter driver deserialization
methods for commands and finders.

Also adds support for the normal defaults to the deserialization
handlers (can revert this if you like, but it helps when interacting
with flutter driver as a client so that it is more similar to how the
API works).

The tests aren't completely exhaustive as that would get very verbose,
but they do cover a variety of use cases.

Fixes https://github.com/flutter/flutter/issues/172127

## Pre-launch Checklist

- [x] I read the [Contributor Guide] and followed the process outlined
there for submitting PRs.
- [x] I read the [Tree Hygiene] wiki page, which explains my
responsibilities.
- [x] I read and followed the [Flutter Style Guide], including [Features
we expect every widget to implement].
- [x] I signed the [CLA].
- [x] I listed at least one issue that this PR fixes in the description
above.
- [x] I updated/added relevant documentation (doc comments with `///`).
- [x] I added new tests to check the change I am making, or this PR is
[test-exempt].
- [x] I followed the [breaking change policy] and added [Data Driven
Fixes] where supported.
- [x] All existing and new tests are passing.