Delete impeller::SPrintF. (#174900)
We have std::format at home.

This is superior for a few reasons. The caller doesn't have to remember
the format string for the type. I believe std::string_views that are not
null terminated are also handled correctly. Also, we don't have to
maintain and test our own SPrintF.